# docs/swagger.yaml
definitions:
  controllers.CreateComments:
    properties:
      message:
        type: string
      photo_id:
        type: integer
    type: object
  controllers.CreatePhotos:
    properties:
      caption:
        type: string
      photo_url:
        type: string
      title:
        type: string
    type: object
...
  controllers.UpdateComments:
    properties:
      message:
        type: string
      photo_id:
        type: integer
    type: object
  controllers.UpdatePhotos:
    properties:
      caption:
        type: string
      photo_url:
        type: string
      title:
        type: string
    type: object
...
  helpers.ErrorResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  models.Comment:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      photo_id:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  models.Photo:
    properties:
      caption:
        type: string
      created_at:
        type: string
      id:
        type: integer
      photo_url:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /comment/:
    get:
      consumes:
      - application/json
      description: Get All Comment
      parameters:
      - description: 'Bearer {access_token} Ex : ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Get All Comment
      tags:
      - Comment
    post:
      consumes:
      - application/json
      description: Create Comment with the given information
      parameters:
      - description: 'Bearer {access_token} Ex : ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment has been created
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/controllers.CreateComments'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Create a new Comment
      tags:
      - Comment
  /comment/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Comment by ID Comment
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Bearer {access_token} Ex : ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Delete Comment
      tags:
      - Comment
    get:
      consumes:
      - application/json
      description: Get One Comment by ID Comment
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Bearer {access_token} Ex : ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Get One Comment
      tags:
      - Comment
    put:
      consumes:
      - application/json
      description: Update Comment by ID Comment
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Bearer {access_token} Ex : ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Comment has been Updated
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdateComments'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Update Comment
      tags:
      - Comment
  /photo/:
    get:
      consumes:
      - application/json
      description: Get All Photo by ID Photo
      parameters:
      - description: 'Bearer {access_token} Ex : ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Get All Photo
      tags:
      - Photo
    post:
      consumes:
      - application/json
      description: Create Photo with the given information
      parameters:
      - description: 'Bearer {access_token} Ex : ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: The social media to create
        in: body
        name: socialmedia
        required: true
        schema:
          $ref: '#/definitions/controllers.CreatePhotos'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Create a new Photo
      tags:
      - Photo
  /photo/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Photo by ID Photo
      parameters:
      - description: Photo ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Bearer {access_token} Ex : ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Delete Photo
      tags:
      - Photo
    get:
      consumes:
      - application/json
      description: Get One Photo by ID Photo
      parameters:
      - description: Photo ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Bearer {access_token} Ex : ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Get One Photo
      tags:
      - Photo
    put:
      consumes:
      - application/json
      description: Update Photo with the given information
      parameters:
      - description: Photo ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Bearer {access_token} Ex : ''Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: The social media to create
        in: body
        name: socialmedia
        required: true
        schema:
          $ref: '#/definitions/controllers.UpdatePhotos'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Photo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.ErrorResponse'
      summary: Update Photo
      tags:
      - Photo
...
swagger: "2.0"